AI-agent/requirements.txt
flask==2.3.3
twilio==8.10.0
cohere==4.37
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
gunicorn==21.2.0
google-generativeai==0.3.2 